refactor: 광고 시간 판별 로직 유틸 함수로 분리
